requirements.txt
telebot
SpeechRecognition
soundfile
moviepy
python-dotenv